Рефакторинг приложения с использованием паттерна MVVM
