--- app/src/main/java/com/example/clubhouse/ui/interfaces/ContactEntityOwner.kt
@@ -1,7 +1,0 @@
-package com.example.clubhouse.ui.interfaces
-
-import com.example.clubhouse.data.ContactEntity
-
-interface ContactEntityOwner {
-    fun getContact(): ContactEntity?
-}
