--- app/src/main/java/com/example/clubhouse/ui/interfaces/ContactServiceConsumer.kt
@@ -1,5 +1,0 @@
-package com.example.clubhouse.ui.interfaces
-
-interface ContactServiceConsumer {
-    fun onServiceBoundListener()
-}
