--- app/src/main/java/com/example/clubhouse/ui/interfaces/ContactServiceOwner.kt
@@ -1,7 +1,0 @@
-package com.example.clubhouse.ui.interfaces
-
-import com.example.clubhouse.ui.services.ContactService
-
-interface ContactServiceOwner {
-    fun getService(): ContactService?
-}
