+++ app/src/main/java/com/example/clubhouse/ui/viewmodels/ContactDetailsViewModel.kt
@@ -1,0 +1,41 @@
+package com.example.clubhouse.ui.viewmodels
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.viewModelScope
+import com.example.clubhouse.data.ContactEntity
+import com.example.clubhouse.data.ContactRepository
+import kotlinx.coroutines.CancellationException
+import kotlinx.coroutines.launch
+import timber.log.Timber
+
+class ContactDetailsViewModel(application: Application) : AndroidViewModel(
+    application
+) {
+    val contact: LiveData<ContactEntity>
+        get() = mutableContact
+
+    private val mutableContact = MutableLiveData<ContactEntity>()
+
+    fun setContactLookup(lookup: String) {
+        if (mutableContact.value != null) {
+            mutableContact.value = mutableContact.value
+        }
+
+        refreshContactDetails(lookup)
+    }
+
+    fun refreshContactDetails(lookup: String) {
+        viewModelScope.launch {
+            try {
+                ContactRepository.getContact(getApplication(), lookup)?.let {
+                    mutableContact.postValue(it)
+                }
+            } catch (e: CancellationException) {
+                Timber.d("ContactDetailsViewModel job cancelled\n$e")
+            }
+        }
+    }
+}
