+++ app/src/main/res/menu/fragment_contact_details_menu.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/menuRefresh"
+        android:title="@string/refresh"
+        app:showAsAction="never" />
+</menu>
